Update integer values to floats and add casts for function return values

+++ src/main.c
@@ -40,17 +40,17 @@ void DrawBall(void);
 void DrawScore(void);
 
 int main(void) {
-    game.screen = (Vector2) {800, 600};
+    game.screen = (Vector2) {800.0f, 600.0f};
 
     InitWindow(game.screen.x, game.screen.y, "Pong");
 
     game.bgColor = WHITE;
     game.halflineColor = GREEN;
 
-    game.ball = (Ball) {.x = game.screen.x / 2, .y = GetRandomValue(10, game.screen.y - 10), .radius = 10.0f, .x_velocity = 5.0f * (!(GetRandomValue(0, 1024) % 2)? 1: -1), .y_velocity = 5.0f * (!(GetRandomValue(0, 1024) % 2)? 1: -1), .color = GREEN};
+    game.ball = (Ball) {.x = game.screen.x / 2.0f, .y = (float) GetRandomValue(10, game.screen.y - 10), .radius = 10.0f, .x_velocity = 5.0f * (!(GetRandomValue(0, 1024) % 2)? 1.0f: -1.0f), .y_velocity = 5.0f * (!(GetRandomValue(0, 1024) % 2)? 1.0f: -1.0f), .color = GREEN};
 
-    game.paddles[0] = (Paddle) {.x = 0.0f, .y = (game.screen.y - game.screen.y / 6) / 2, .width = 10.0f, .height = game.screen.y / 6, .velocity = 10.0f, .color = GREEN};
-    game.paddles[1] = (Paddle) {.x = game.screen.x - 10.0f, .y = (game.screen.y - game.screen.y / 6) / 2, .width = 10.0f, .height = game.screen.y / 6, .velocity = 10.0f, .color = GREEN};
+    game.paddles[0] = (Paddle) {.x = 0.0f, .y = (game.screen.y - game.screen.y / 6.0f) / 2.0f, .width = 10.0f, .height = game.screen.y / 6.0f, .velocity = 10.0f, .color = GREEN};
+    game.paddles[1] = (Paddle) {.x = game.screen.x - 10.0f, .y = (game.screen.y - game.screen.y / 6.0f) / 2.0f, .width = 10.0f, .height = game.screen.y / 6.0f, .velocity = 10.0f, .color = GREEN};
 
     game.score = (Score) {.p1 = 0, .p2 = 0, .size = 30, .color = GREEN};
 
@@ -95,8 +95,8 @@ void UpdateDrawFrame(void) {
 }
 
 void DrawHalfline(void) {
-    for (int i = 0; i <= game.screen.y; i += game.screen.y / 20) {
-        DrawLine(game.screen.x / 2, i, game.screen.x / 2, i + game.screen.y / 40, game.halflineColor);
+    for (float i = 0.0f; i <= game.screen.y; i += game.screen.y / 20.0f) {
+        DrawLine(game.screen.x / 2.0f, i, game.screen.x / 2.0f, i + game.screen.y / 40.0f, game.halflineColor);
     }
 }
 
@@ -112,7 +112,7 @@ void MovePaddles(void) {
         game.paddles[1].y += game.paddles[1].velocity;
 
     for (int i = 0; i < 2; i++) {
-        if (game.paddles[i].y < 0) game.paddles[i].y = 0.0f;
+        if (game.paddles[i].y < 0.0f) game.paddles[i].y = 0.0f;
         if (game.paddles[i].y > (game.screen.y - game.paddles[i].height)) game.paddles[i].y = game.screen.y - game.paddles[i].height;
     }
 }
@@ -133,24 +133,24 @@ void MoveBall(void) {
     if (game.ball.y > game.screen.y - game.ball.radius) game.ball.y = game.screen.y - game.ball.radius;
 
     if ((game.ball.y + game.ball.radius) >= game.screen.y || (game.ball.y - game.ball.radius) <= 0.0f)
-        game.ball.y_velocity *= -1;
+        game.ball.y_velocity *= -1.0f;
 
     if (CheckCollisionCircleRec((Vector2) {game.ball.x, game.ball.y}, game.ball.radius, (Rectangle) {game.paddles[0].x, game.paddles[0].y, game.paddles[0].width, game.paddles[0].height}) || CheckCollisionCircleRec((Vector2) {game.ball.x, game.ball.y}, game.ball.radius, (Rectangle) {game.paddles[1].x, game.paddles[1].y, game.paddles[1].width, game.paddles[1].height})) {
-        game.ball.x_velocity *= -1;
+        game.ball.x_velocity *= -1.0f;
     }
 
     if ((game.ball.x - game.ball.radius) <= 0.0f) {
         game.score.p2++;
-        game.ball.x = game.screen.x / 2;
-        game.ball.y = GetRandomValue(game.ball.radius, game.screen.y - game.ball.radius);
-        game.ball.y_velocity *= !(GetRandomValue(0, 1024) % 2)? 1: -1;
+        game.ball.x = game.screen.x / 2.0f;
+        game.ball.y = (float) GetRandomValue(game.ball.radius, game.screen.y - game.ball.radius);
+        game.ball.y_velocity *= !(GetRandomValue(0, 1024) % 2)? 1.0f: -1.0f ;
     }
 
     if ((game.ball.x + game.ball.radius) >= game.screen.x) {
         game.score.p1++;
-        game.ball.x = game.screen.x / 2;
-        game.ball.y = GetRandomValue(game.ball.radius, game.screen.y - game.ball.radius);
-        game.ball.y_velocity *= !(GetRandomValue(0, 1024) % 2)? 1: -1;
+        game.ball.x = game.screen.x / 2.0f;
+        game.ball.y = (float) GetRandomValue(game.ball.radius, game.screen.y - game.ball.radius);
+        game.ball.y_velocity *= !(GetRandomValue(0, 1024) % 2)? 1.0f: -1.0f;
     }
 }
 
@@ -161,5 +161,5 @@ void DrawBall(void) {
 void DrawScore(void) {
     const char *text = TextFormat("%d %d", game.score.p1, game.score.p2);
     int width = MeasureText(text, game.score.size);
-    DrawText(text, (game.screen.x - width) / 2, 0.0f, game.score.size, game.score.color);
+    DrawText(text, (game.screen.x - width) / 2.0f, 0.0f, game.score.size, game.score.color);
 }
